requirements.txt
absl-py==0.15.0
astunparse==1.6.3
clang==5.0
flatbuffers==1.12
gast==0.4.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.1.0
importlib_metadata==8.6.1
joblib==1.4.2
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.8
MarkupSafe==3.0.2
numpy==1.19.5
opt-einsum==3.3.0
packaging==25.0
pandas==1.4.4
protobuf==3.20.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.15.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-cpu==2.6.0
tensorflow-estimator==2.15.0
termcolor==1.1.0
threadpoolctl==3.6.0
typing-extensions==3.7.4.3
Werkzeug==3.1.3
wrapt==1.12.1
zipp==3.21.0
flask